Review changes
Move QPCR instruction out of wtype, some clean up of code and import directives.
