+++ antha/ast/qpcrinstruction.go
@@ -1,0 +1,27 @@
+package ast
+
+import (
+	"fmt"
+	"github.com/antha-lang/antha/antha/anthalib/wtype"
+)
+
+// QPCRInstruction is a high-level instruction to perform a QPCR analysis.
+type QPCRInstruction struct {
+	ID           string
+	ComponentIn  *wtype.LHComponent
+	ComponentOut *wtype.LHComponent
+	Definition   string
+	Barcode      string
+	Command      string
+}
+
+func (ins QPCRInstruction) String() string {
+	return fmt.Sprint("QPCRInstruction")
+}
+
+// NewQPCRInstruction creates a new QPCRInstruction
+func NewQPCRInstruction() *QPCRInstruction {
+	var inst QPCRInstruction
+	inst.ID = wtype.GetUUID()
+	return &inst
+}
